Make createAppUser private
