+++ src/definition/accessors/index.ts
@@ -42,7 +42,6 @@ import { ISlashCommandsExtend } from './ISlashCommandsExtend';
 import { ISlashCommandsModify } from './ISlashCommandsModify';
 import { IUploadRead } from './IUploadRead';
 import { IUserBuilder } from './IUserBuilder';
-import { IUserCreator } from './IUserCreator';
 import { IUserRead } from './IUserRead';
 
 export {
@@ -85,7 +84,6 @@ export {
     ISlashCommandsExtend,
     ISlashCommandsModify,
     IUserBuilder,
-    IUserCreator,
     IUploadRead,
     IUserRead,
     LogMessageSeverity,
