+++ src/server/accessors/index.ts
@@ -28,7 +28,6 @@ import { SlashCommandsExtend } from './SlashCommandsExtend';
 import { SlashCommandsModify } from './SlashCommandsModify';
 import { UploadRead } from './UploadRead';
 import { UserBuilder } from './UserBuilder';
-import { UserCreator } from './UserCreator';
 import { UserRead } from './UserRead';
 
 export {
@@ -62,6 +61,5 @@ export {
     SlashCommandsModify,
     UploadRead,
     UserBuilder,
-    UserCreator,
     UserRead,
 };
