+++ tests/server/accessors/UserBuilder.spec.ts
@@ -2,13 +2,11 @@ import { Expect, Test } from 'alsatian';
 
 import { IUserCreation } from '../../../src/definition/users';
 import { UserBuilder } from '../../../src/server/accessors';
-import { TestData } from '../../test-data/utilities';
 
 export class UserBuilderAccessorTestFixture {
     @Test()
     public basicUserBuilder() {
         Expect(() => new UserBuilder()).not.toThrow();
-        Expect(() => new UserBuilder(TestData.getUserCreator())).not.toThrow();
     }
 
     @Test()
